refactoring part one
