+++ tsconfig.json
@@ -7,6 +7,7 @@
         "./types/",
     ],
     "target": "esnext",
+
     "strictPropertyInitialization": false,
     "module": "esnext",
     "strict": true,
